AliExpress/requirements.txt
async-generator==1.10
attrs==22.1.0
Automat==20.2.0
beautifulsoup4==4.11.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
docker==5.0.3
docopt==0.6.2
filelock==3.8.0
firebase-admin==5.3.0
google-api-core==2.10.0
google-api-python-client==2.60.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpcio==1.48.1
grpcio-status==1.48.1
h11==0.13.0
herokuify-scrapyd==1.0
htmldocx==0.0.6
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.7.0
itemloaders==1.0.4
jmespath==1.0.1
lxml==4.9.1
msgpack==1.0.4
outcome==1.2.0
parsel==1.6.0
pipreqs==0.4.11
Protego==0.2.1
proto-plus==1.22.1
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-docx==0.8.11
python-dotenv==0.20.0
pywin32==227
PyYAML==6.0
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
retrying==1.3.3
rsa==4.9
scrapinghub==2.4.0
Scrapy==2.6.2
scrapyd==1.3.0
scrapyd-client==1.2.2
selenium==4.4.3
service-identity==21.1.0
shub==2.14.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
tldextract==3.3.1
toml==0.10.2
tqdm==4.55.1
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typing_extensions==4.3.0
uberegg==0.1.1
uritemplate==4.1.1
urllib3==1.26.11
w3lib==2.0.1
webdriver-manager==3.8.3
websocket-client==1.3.3
wsproto==1.1.0
yarg==0.1.9
zope.interface==5.4.0
